+++ src/main/java/spring/backend/member/application/ReadMemberHomeService.java
@@ -5,9 +5,9 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spring.backend.activity.dto.response.HomeActivityInfoResponse;
-import spring.backend.activity.dto.response.QuickStartResponse;
+import spring.backend.quickstart.dto.response.QuickStartResponse;
 import spring.backend.activity.query.dao.ActivityDao;
-import spring.backend.activity.query.dao.QuickStartDao;
+import spring.backend.quickstart.query.dao.QuickStartDao;
 import spring.backend.member.domain.entity.Member;
 import spring.backend.member.dto.response.HomeMainResponse;
 import spring.backend.member.dto.response.HomeMemberInfoResponse;
