+++ src/main/java/spring/backend/member/presentation/ValidateNicknameController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.core.presentation.RestResponse;
-import spring.backend.member.application.ValidateNicknameService;
+import spring.backend.member.domain.service.ValidateNicknameService;
 import spring.backend.member.presentation.swagger.ValidateNicknameSwagger;
 
 @RestController
